Fix CLI import
